requirements.txt
flask
flask-migrate
flask-wtf
flask_sqlalchemy
flask_migrate
python-dotenv
flask-login
email_validator
flask-mail
pyjwt
flask-bootstrap
flask-moment
flask-babel
langdetect
requests
opensearch-py
rq
httpie
flask-httpauth